Working on making it multi-distro aware.
Management side of the house seems to install well, now working on
client.

Had a lot of issues getting MongoDB to stand up, but it turns out there
was issues with permissions on the /var/run directory. Put a play in
there to solve that and we were off to the races.

More to come on the client side.